LayoutTests/fast/dom/dom-parser-head-body-order-expected.txt

This tests that head appears before body in the result of (new DOMParser).parseFromString

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS doc.documentElement.children[0] is doc.head
PASS doc.documentElement.children[1] is doc.body
PASS successfullyParsed is true

TEST COMPLETE
